#4 [separate] separate class and package
