+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/api/GithubApi.kt
@@ -1,9 +1,7 @@
 package jp.co.yumemi.android.code_check.api
 
-import jp.co.yumemi.android.code_check.model.ApiResult
-import retrofit2.Response
+import jp.co.yumemi.android.code_check.model.dto.ApiResult
 import retrofit2.http.GET
-import retrofit2.http.Headers
 import retrofit2.http.Query
 
 interface GithubApi {
