+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/model/DetailItem.kt
@@ -1,0 +1,16 @@
+package jp.co.yumemi.android.code_check.model
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+
+@Parcelize
+data class DetailItem(
+    val fullName: String,
+    val avatarUrl: String,
+    val language: String?,
+    val stargazersCount: Long,
+    val watchersCount: Long,
+    val forksCount: Long,
+    val openIssuesCount: Long,
+) : Parcelable
