+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/model/dto/ApiResult.kt
@@ -1,9 +1,9 @@
-package jp.co.yumemi.android.code_check.model
+package jp.co.yumemi.android.code_check.model.dto
 
 
 import android.util.Log
 import com.google.gson.annotations.SerializedName
-import jp.co.yumemi.android.code_check.DetailItem
+import jp.co.yumemi.android.code_check.model.DetailItem
 
 data class ApiResult(
     @SerializedName("items")
