+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/model/dto/Item.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.code_check.model
+package jp.co.yumemi.android.code_check.model.dto
 
 
 import com.google.gson.annotations.SerializedName
@@ -20,4 +20,4 @@ data class Item(
     @SerializedName("open_issues_count")
     val openIssuesCount: Int,
 
-)
+    )
