+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/MainActivity.kt
